+++ CHANGELOG.md
@@ -6,6 +6,12 @@ This is the changelog for SpotBugs. This follows [Keep a Changelog v1.0.0](http:
 Currently the versioning policy of this project follows [Semantic Versioning v2.0.0](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased - 2025-??-??
+### Fixed
+- Fix for missing -adjustPriority parameter in Eclipse preferences ([#3687](https://github.com/spotbugs/spotbugs/issues/3687))
+- Documentation of -adjustPriority parameter
+
+### Changed
+- Support for fully qualified class names for detectors in -adjustPriority parameter
 
 ## 4.9.6 - 2025-09-16
 ### Fixed